requirements.txt
torch
python-dateutil
click
tqdm
gitpython
lxml
doxmlparser